\name{Adoptees}
\docType{data}
\alias{Adoptees}
\title{Adoptees Education }
\description{
  a cross-section

 \emph{number of observations} :  16481

   \emph{country} :  United States

\emph{economic topic} : labor economics

\emph{econometrics topic} : binomial model

}
\usage{data(Adoptees)}
\format{A dataframe containing :
  \describe{
    \item{adopt}{is the child adopted ?}
    \item{educ}{years of education}
    \item{coll}{college graduate}
    \item{stinsch}{ still in school}
    \item{gender}{one of 'male' and 'female'}
    \item{age}{the age in years}
    \item{feduc}{father's education}
    \item{meduc}{mother's education}
    \item{fcoll}{is the father college graduate ?}
    \item{mcoll}{is the mother college graduate ?}
    \item{linc}{log of income}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Plug, Erik (2004) \dQuote{Estimating the effect of mother's schooling on children's schooling using a sample of adoptees}, \emph{American Economic Review}, \bold{94(1)}, 358-368.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
