\name{AirlinesExCap}
\docType{data}
\alias{AirlinesExCap}
\title{Excess Capacity in Us Airlines }
\description{
  yearly observations of  22 firms from 1971  to 1986

 \emph{number of observations} :  268

   \emph{country} :  USA

\emph{economic topic} : producer behavior

\emph{econometrics topic} : error component

}
\usage{data(AirlinesExCap)}
\format{A dataframe containing :
  \describe{
    \item{firm}{the firm}
    \item{year}{the year}
    \item{cost}{the variable cost}
    \item{output}{the output, a multilateral index of actual seat miles flown}
    \item{capital}{capital index, passenger seating and cargo tonnage capacity available for use}
    \item{pl}{wage index}
    \item{pm}{material price index}
    \item{pf}{fuel price index}
    \item{pk}{user cost of capital}
    \item{sl}{the share of labor cost}
    \item{sm}{the share of material cost}
    \item{revshare}{revenu share}
    \item{stage}{stage length, average length in miles of a firm's flights}
  }
}
\source{
    Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.

}
\references{
    Badi H. Baltagi, James M. Griffin (1998) \dQuote{Excess Capacity: A Permanent Characteristic of U.S. Airlines}, \emph{Journal of Applied Econometrics}, \bold{13(5)}, 645--657.

    H. Baltagi, James M. Griffin (1995) \dQuote{Airline deregulation : the cost pieces of the puzzle}, \emph{International Economic Review}, \bold{36}, 245--259.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
