\name{Apples}
\docType{data}
\alias{Apples}
\title{Apple Production }
\description{
  yearly observations of  173 farms from 1984  to 1986

 \emph{number of observations} :  405

   \emph{country} :  France

\emph{economic topic} : producer behavior

\emph{econometrics topic} : seemingly unrelated regression

}
\usage{data(Apples)}
\format{A dataframe containing :
  \describe{
    \item{id}{farm's id}
    \item{year}{year}
    \item{capital}{capital stock}
    \item{labor}{quantity of labor}
    \item{materials}{quantity of materials}
    \item{apples}{production of apples}
    \item{otherprod}{other productions}
    \item{pc}{price of capital}
    \item{pl}{price of labor}
    \item{pm}{price of materials}
  }
}
\source{
    Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.

}
\references{
    Marc Ivaldi, Norbert Ladoux (1996) \dQuote{Comparing Fourier and Translog Specifications of Multiproduct Technology: Evidence from an Incomplete Panel of French Farmers}, \emph{Journal of Applied Econometrics}, \bold{11(6)}, 649--668.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
