\name{CallBacks}
\docType{data}
\alias{CallBacks}
\title{Callbacks to Job Applications }
\description{
  a cross-section

 \emph{number of observations} :  6072

   \emph{country} :  United States

\emph{economic topic} : labor and demographic economics

\emph{econometrics topic} : binomial variable

}
\usage{data(CallBacks)}
\format{A dataframe containing :
  \describe{
    \item{jobid}{the job index}
    \item{unempdur}{unemployment duration in month}
    \item{interim}{a dummy for interim experience}
    \item{callback}{a dummy for call backs}
    \item{old}{a dummy for age 57-58}
  }
}
\source{
    American Economic Association's website \url{http://www.aeaweb.org}.

}
\references{
    Farber, Henry S.,  Dan   Silverman and  Till  von Wachter (2016) \dQuote{Determinants of callbacks to job applications: an audit study}, \emph{American Economic Review}, \bold{106(5)}, 314-318.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
