\name{CoordFailure}
\docType{data}
\alias{CoordFailure}
\title{How to Overcome Organization Failure in Organization }
\description{
  a cross-section

 \emph{number of observations} :  7168

   \emph{country} :  United States and Spain

\emph{economic topic} : labor and demographic economics

\emph{econometrics topic} : binomial variable

}
\usage{data(CoordFailure)}
\format{A dataframe containing :
  \describe{
    \item{firm}{the firm index}
    \item{id}{ the individual index}
    \item{period}{the period}
    \item{place}{either Cleveland or Barcelona}
    \item{bonus1}{the bonus for the first block of 10 rounds}
    \item{bonus2}{the bonus for the second block of 10 rounds}
    \item{bonus3}{the bonus for the third block of 10 rounds}
    \item{effort}{the level of effort of the employee}
  }
}
\source{
    American Economic Association's website \url{http://www.aeaweb.org}.

}
\references{
    Brandts, Jordi  and  David J.  Cooper (2006) \dQuote{A change would do you good... An experimental study on how to overcome coordination failure in organizations}, \emph{American Economic Review}, \bold{96(3)}, 669-693.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
