\name{etw}
\docType{data}
\alias{etw}
\title{Spatial weights matrix for EvapoTransp}

\description{
  Spatial weights matrix for the EvapoTransp data frame

}

\usage{
data(etw)
}

\format{ A 86x86 matrix with elements different from zero if area i and
  j are neighbours. Weights are row standardized.  }

\author{Giovanni Millo}

\keyword{datasets}
