\name{EvapoTransp}
\docType{data}
\alias{EvapoTransp}
\title{Evapotranspiration }
\description{
  a cross-section

 \emph{number of observations} :  430

   \emph{country} :  France

\emph{economic topic} : agricultural and natural resource economics, environmental and ecological economics

\emph{econometrics topic} : spatial data

}
\usage{data(EvapoTransp)}
\format{A dataframe containing :
  \describe{
    \item{id}{observation site}
    \item{period}{measuring period}
    \item{et}{evapotranspiration}
    \item{prec}{precipitation}
    \item{meansmd}{mean soil moisture deficit}
    \item{potet}{potential evapotranspiration}
    \item{infil}{infiltration rate}
    \item{biomass}{biomass}
    \item{biomassp1}{biomass in early growing season}
    \item{biomassp2}{biomass in main growth period}
    \item{biomassp3}{peak biomass}
    \item{biomassp4}{peak biomass after clipping}
    \item{biomassp5}{biomass in autumn}
    \item{plantcover}{plant cover}
    \item{softforbs}{soft-leaved forbs}
    \item{tallgrass}{tall grass}
    \item{diversity}{species diversity}
    \item{matgram}{mat-forming graminoids}
    \item{dwarfshrubs}{dwarf shrubs}
    \item{legumes}{abundance of legumes}
  }
}
\source{
    Authors.

}
\references{
    Obojes, N.,  M.   Bahn,   E.   Tasser,   J.   Walde,  N.   Inauen,  E.   Hiltbrunner,  P.   Saccone,  J.   Lochet,  J.   Cl\'ement and  S.   Lavorel () \dQuote{Vegetation effects on the water balance of mountain grasslands depend on climatic conditions}, \emph{Ecohydrology}, \bold{8(4)}, 552-569.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
