\name{GiantsShoulders}
\docType{data}
\alias{GiantsShoulders}
\title{Impact of Institutions on Cumulative Research }
\description{
  yearly observations of  216 articles from 1970  to 2001

 \emph{number of observations} :  4880

   \emph{country} :  United States

\emph{economic topic} : economic development, innovation, technological change, and growth

\emph{econometrics topic} : count data

}
\usage{data(GiantsShoulders)}
\format{A dataframe containing :
  \describe{
    \item{pair}{the pair article index}
    \item{article}{the article index}
    \item{brc}{material of the article is deposit on a Biological Ressource Center}
    \item{pubyear}{publication year of the article}
    \item{brcyear}{year of the deposit in brc of the material related to the article}
    \item{year}{the year index}
    \item{citations}{the number of citations}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Furman, Jeffrey L.  and  Scott  Stern (2011) \dQuote{Climbing atop the shoulders of giants: the impact of institutions on cumulative research}, \emph{American Economic Review}, \bold{101(5)}, 1933-1963.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
