\name{IncomeMigrationH}
\docType{data}
\alias{IncomeMigrationH}
\title{Income and Migration, Household Data}
\description{yearly observations of 317 households from 2000 to 2006

    \emph{number of observations} : 2219

    \emph{number of  time-series } :  7 

    \emph{country} :  Indonesia 

    \emph{package} : limdeppanel

    \emph{JEL codes: }F22, J43, O13, O15, Q11, Q12, R23

    \emph{Chapter} : 08

}
\usage{data(IncomeMigrationH)}
\format{A dataframe containing:
  \describe{
    \item{household}{household index}
    \item{year}{the year}
    \item{migration}{a dummy indicating whether a household has any migrant departing in year t+1}
    \item{price}{rice price shock}
    \item{rain}{rain shock}
    \item{land}{landholdings (ha)}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Bazzi, Samuel (2017) \dQuote{Wealth Heterogeneity and the Income Elasticity of Migration}, \emph{American Economic Journal, Applied Economics}, \bold{9(2)}, 219--255, \doi{10.1257/app.20150548}
.

}
\keyword{datasets}
