\name{IncomeMigrationV}
\docType{data}
\alias{IncomeMigrationV}
\title{Income and Migration, Village Data}
\description{3-yearly observations of 44674 villages from 2005 to 2008

    \emph{number of observations} : 89348

    \emph{number of  time-series } :  2 

    \emph{country} :  Indonesia 

    \emph{package} : panellimdep

    \emph{JEL codes: }F22, J43, O13, O15, Q11, Q12, R23

    \emph{Chapter} : 08

}
\usage{data(IncomeMigrationV)}
\format{A dataframe containing:
  \describe{
    \item{village}{village index}
    \item{year}{the year}
    \item{emigration}{share of the emigrants in the total population}
    \item{district}{the district of the village}
    \item{price}{rice price shock}
    \item{rain}{rain shock}
    \item{pareto}{Pareto parameter of the landholdings distribution}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Bazzi, Samuel (2017) \dQuote{Wealth Heterogeneity and the Income Elasticity of Migration}, \emph{American Economic Journal, Applied Economics}, \bold{9(2)}, 219--255, \doi{10.1257/app.20150548}
.

}
\keyword{datasets}
