\name{Index.chapter}
\docType{data}
\alias{Index.Chapter}
\title{Chapter}
\description{
\itemize{
\item \bold{ 01 } :  Introduction 
\itemize{
\item \code{\link{Tileries}} : Production of tileries in Egypt
}
\item \bold{ 02 } :  The error component model 
\itemize{
\item \code{\link{DemocracyIncome}} : The relation between democraty and income
\item \code{\link{DemocracyIncome25}} : The relation between democraty and income
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing countries
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{TobinQ}} : The Q theory of investment
\item \code{\link{TurkishBanks}} : Turkish Banks
}
\item \bold{ 03 } :  Advanced error components models 
\itemize{
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
}
\item \bold{ 04 } :  Tests on error components models 
\itemize{
\item \code{\link{RDSpillovers}} : Research and development spillovers data
}
\item \bold{ 05 } :  Robust inference and estimation 
\itemize{
\item \code{\link{RDSpillovers}} : Research and development spillovers data
}
\item \bold{ 06 } :  Endogeneity 
\itemize{
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing countries
\item \code{\link{Mafia}} : Mafia and Public Spending
\item \code{\link{SeatBelt}} : Seat belt usage and traffic fatalities
\item \code{\link{TradeEU}} : Trade in the European Union
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ 07 } :  Estimation of a dynamic model 
\itemize{
\item \code{\link{DemocracyIncome}} : The relation between democraty and income
\item \code{\link{DemocracyIncome25}} : The relation between democraty and income
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IneqGrowth}} : Inequality and growth
\item \code{\link{RDPerfComp}} : R and D performing companies
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
\item \code{\link{Solow}} : Growth model
}
\item \bold{ 08 } :  Count data and limited dependent variables 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
\item \code{\link{CoordFailure}} : How to overcome organization failure in organization
\item \code{\link{Donors}} : Dynamics of charitable giving
\item \code{\link{GiantsShoulders}} : Impact of institutions on cumulative research
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
\item \code{\link{LandReform}} : Politics and land reforms in India
\item \code{\link{LateBudgets}} : Late Budgets
\item \code{\link{MagazinePrices}} : Magazine prices
\item \code{\link{Reelection}} : Deficits and reelection
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ 09 } :  Panel time series 
\itemize{
\item \code{\link{Dialysis}} : Diffusion of haemodialysis technology
\item \code{\link{HousePricesUS}} : House Prices data
\item \code{\link{RDSpillovers}} : Research and development spillovers data
}
\item \bold{ 10 } :  Spatial panels 
\itemize{
\item \code{\link{EvapoTransp}} : Evapotranspiration
\item \code{\link{HousePricesUS}} : House Prices data
}
}
}
\keyword{documentation}
