\name{Index.Economics}
\docType{data}
\alias{Index.Economics}
\title{Economic fields}
\description{
\itemize{
\item agricultural and natural resource economics, environmental and ecological economics 
 \itemize{
\item \code{\link{EvapoTransp}} : Evapotranspiration  
\item \code{\link{LandReform}} : Politics and Land Reforms in India  
}
\item economic development, innovation, technological change, and growth 
 \itemize{
\item \code{\link{FinanceGrowth}} : Financial Institutions and Growth  
\item \code{\link{GiantsShoulders}} : Impact of Institutions on Cumulative Research  
\item \code{\link{IncomeMigrationH}} : Income and Migration (household Data)  
\item \code{\link{IncomeMigrationV}} : Income and Migration (village Data)  
\item \code{\link{IneqGrowth}} : Inequality and Growth  
\item \code{\link{RDPerfComp}} : R and D Performing Companies  
\item \code{\link{RDSpillovers}} : Research and Development Spillovers Data  
\item \code{\link{ScrambleAfrica}} : the Long-run Effects of the Scramble for Africa  
\item \code{\link{Solow}} : Growth Model  
}
\item financial economics 
 \itemize{
\item \code{\link{SovereignDefault}} : Capital Flows, Commodities and Sovereign Defaults  
\item \code{\link{TwinCrises}} : Costs of Currency and Banking Crises  
}
\item industrial organization 
 \itemize{
\item \code{\link{MagazinePrices}} : Magazine Prices  
}
\item international economics 
 \itemize{
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing Countries  
\item \code{\link{TradeEU}} : Trade in the European Union  
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States  
}
\item labor and demographic economics 
 \itemize{
\item \code{\link{CallBacks}} : Callbacks to Job Applications  
\item \code{\link{CoordFailure}} : How to Overcome Organization Failure in Organization  
\item \code{\link{Seniors}} : Intergenerationals Experiments  
}
\item law and economics 
 \itemize{
\item \code{\link{SeatBelt}} : Seat Belt Usage and Traffic Fatalities  
}
\item macroeconomics and monetary policy 
 \itemize{
\item \code{\link{Mafia}} : Mafia and Public Spending  
}
\item microeconomics 
 \itemize{
\item \code{\link{Donnors}} : Dynamics of Charitable Giving  
\item \code{\link{InvRate}} : Investment Rate  
\item \code{\link{TexasElectr}} : Production of Electricity in Texas  
\item \code{\link{Tileries}} : Production of Tileries in Egypt  
\item \code{\link{TobinQ}} : the Q Theory of Investment  
\item \code{\link{TurkishBanks}} : Turkish Banks  
}
\item public economics 
 \itemize{
\item \code{\link{DemocracyIncome}} : the Relation Between Democraty and Income  
\item \code{\link{DemocracyIncome25}} : the Relation Between Democraty and Income  
\item \code{\link{Dialysis}} : Diffusion of Haemodialysis Technology  
\item \code{\link{LateBudgets}} : Late Budgets  
\item \code{\link{Reelection}} : Deficits and Reelection  
\item \code{\link{RegIneq}} : Interregional Redistribution and Inequalities  
}
\item urban, rural, regional, real estate and transportation economics 
 \itemize{
\item \code{\link{HousePricesUS}} : House Prices Data  
}
}

}
\keyword{documentation}
