\name{Index.jel}
\docType{data}
\alias{Index.JEL codes}
\title{JEL codes}
\description{
\itemize{
\item \bold{ C13 } :  Estimation: General 
\itemize{
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
}
\item \bold{ C23 } :  Single Equation Models; Single Variables: Panel Data Models; Spatio-temporal Models 
\itemize{
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IneqGrowth}} : Inequality and growth
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
}
\item \bold{ C33 } :  Multiple or Simultaneous Equation Models: Panel Data Models; Spatio-temporal Models 
\itemize{
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IneqGrowth}} : Inequality and growth
}
\item \bold{ C51 } :  Model Construction and Estimation 
\itemize{
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing countries
\item \code{\link{HousePricesUS}} : House Prices data
\item \code{\link{RDPerfComp}} : R and D performing companies
\item \code{\link{RDSpillovers}} : Research and development spillovers data
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
\item \code{\link{TradeEU}} : Trade in the European Union
}
\item \bold{ C78 } :  Bargaining Theory; Matching Theory 
\itemize{
\item \code{\link{LateBudgets}} : Late Budgets
}
\item \bold{ C90 } :  Design of Experiments: General 
\itemize{
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ C92 } :  Design of Experiments: Laboratory, Group Behavior 
\itemize{
\item \code{\link{CoordFailure}} : How to overcome organization failure in organization
}
\item \bold{ C93 } :  Field Experiments 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ D02 } :  Institutions: Design, Formation, Operations, and Impact 
\itemize{
\item \code{\link{GiantsShoulders}} : Impact of institutions on cumulative research
}
\item \bold{ D23 } :  Organizational Behavior; Transaction Costs; Property Rights 
\itemize{
\item \code{\link{CoordFailure}} : How to overcome organization failure in organization
}
\item \bold{ D24 } :  Production; Cost; Capital; Capital, Total Factor, and Multifactor Productivity; Capacity 
\itemize{
\item \code{\link{RDPerfComp}} : R and D performing companies
\item \code{\link{RDSpillovers}} : Research and development spillovers data
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
\item \code{\link{TurkishBanks}} : Turkish Banks
}
\item \bold{ D64 } :  Altruism; Philanthropy; Intergenerational Transfers 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ D72 } :  Political Processes: Rent-seeking, Lobbying, Elections, Legislatures, and Voting Behavior 
\itemize{
\item \code{\link{DemocracyIncome}} : The relation between democraty and income
\item \code{\link{DemocracyIncome25}} : The relation between democraty and income
\item \code{\link{LandReform}} : Politics and land reforms in India
\item \code{\link{LateBudgets}} : Late Budgets
\item \code{\link{Mafia}} : Mafia and Public Spending
\item \code{\link{Reelection}} : Deficits and reelection
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ D74 } :  Conflict; Conflict Resolution; Alliances; Revolutions 
\itemize{
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ D82 } :  Asymmetric and Private Information; Mechanism Design 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ D83 } :  Search; Learning; Information and Knowledge; Communication; Belief; Unawareness 
\itemize{
\item \code{\link{GiantsShoulders}} : Impact of institutions on cumulative research
}
\item \bold{ E24 } :  Employment; Unemployment; Wages; Intergenerational Income Distribution; Aggregate Human Capital; Aggregate Labor Productivity 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
}
\item \bold{ E32 } :  Business Fluctuations; Cycles 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
}
\item \bold{ E62 } :  Fiscal Policy 
\itemize{
\item \code{\link{Mafia}} : Mafia and Public Spending
\item \code{\link{Reelection}} : Deficits and reelection
}
\item \bold{ F12 } :  Models of Trade with Imperfect Competition and Scale Economies; Fragmentation 
\itemize{
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States
}
\item \bold{ F14 } :  Empirical Studies of Trade 
\itemize{
\item \code{\link{TradeEU}} : Trade in the European Union
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States
}
\item \bold{ F17 } :  Trade: Forecasting and Simulation 
\itemize{
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing countries
}
\item \bold{ F21 } :  International Investment; Long-term Capital Movements 
\itemize{
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States
}
\item \bold{ F22 } :  International Migration 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
}
\item \bold{ F23 } :  Multinational Firms; International Business 
\itemize{
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States
}
\item \bold{ F32 } :  Current Account Adjustment; Short-term Capital Movements 
\itemize{
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ F51 } :  International Conflicts; Negotiations; Sanctions 
\itemize{
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ G15 } :  International Financial Markets 
\itemize{
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ G20 } :  Financial Institutions and Services: General 
\itemize{
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
}
\item \bold{ G21 } :  Banks; Depository Institutions; Micro Finance Institutions; Mortgages 
\itemize{
\item \code{\link{TurkishBanks}} : Turkish Banks
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ H23 } :  Taxation and Subsidies: Externalities; Redistributive Effects; Environmental Taxes and Subsidies 
\itemize{
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ H41 } :  Public Goods 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ H61 } :  National Budget; Budget Systems 
\itemize{
\item \code{\link{LateBudgets}} : Late Budgets
}
\item \bold{ H62 } :  National Deficit; Surplus 
\itemize{
\item \code{\link{Reelection}} : Deficits and reelection
}
\item \bold{ H71 } :  State and Local Taxation, Subsidies, and Revenue 
\itemize{
\item \code{\link{Mafia}} : Mafia and Public Spending
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ H72 } :  State and Local Budget and Expenditures 
\itemize{
\item \code{\link{LateBudgets}} : Late Budgets
}
\item \bold{ H73 } :  State and Local Government; Intergovernmental Relations: Interjurisdictional Differentials and Their Effects 
\itemize{
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ H77 } :  Intergovernmental Relations; Federalism; Secession 
\itemize{
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ I18 } :  Health: Government Policy; Regulation; Public Health 
\itemize{
\item \code{\link{Dialysis}} : Diffusion of haemodialysis technology
}
\item \bold{ I23 } :  Higher Education; Research Institutions 
\itemize{
\item \code{\link{GiantsShoulders}} : Impact of institutions on cumulative research
}
\item \bold{ J14 } :  Economics of the Elderly; Economics of the Handicapped; Non-labor Market Discrimination 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ J15 } :  Economics of Minorities, Races, Indigenous Peoples, and Immigrants; Non-labor Discrimination 
\itemize{
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ J22 } :  Time Allocation and Labor Supply 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
}
\item \bold{ J23 } :  Labor Demand 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
}
\item \bold{ J26 } :  Retirement; Retirement Policies 
\itemize{
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ J31 } :  Wage Level and Structure; Wage Differentials 
\itemize{
\item \code{\link{TexasElectr}} : Production of electricity in Texas
\item \code{\link{Tileries}} : Production of tileries in Egypt
}
\item \bold{ J43 } :  Agricultural Labor Markets 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
}
\item \bold{ J64 } :  Unemployment: Models, Duration, Incidence, and Job Search 
\itemize{
\item \code{\link{CallBacks}} : Callbacks to job applications
}
\item \bold{ K42 } :  Illegal Behavior and the Enforcement of Law 
\itemize{
\item \code{\link{Mafia}} : Mafia and Public Spending
\item \code{\link{SeatBelt}} : Seat belt usage and traffic fatalities
}
\item \bold{ L31 } :  Nonprofit Institutions; NGOs; Social Entrepreneurship 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ L33 } :  Comparison of Public and Private Enterprises and Nonprofit Institutions; Privatization; Contracting Out 
\itemize{
\item \code{\link{TurkishBanks}} : Turkish Banks
}
\item \bold{ L82 } :  Entertainment; Media 
\itemize{
\item \code{\link{MagazinePrices}} : Magazine prices
}
\item \bold{ M12 } :  Personnel Management; Executives; Executive Compensation 
\itemize{
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ M51 } :  Personnel Economics: Firm Employment Decisions; Promotions 
\itemize{
\item \code{\link{Seniors}} : Intergenerationals experiments
}
\item \bold{ O13 } :  Economic Development: Agriculture; Natural Resources; Energy; Environment; Other Primary Products 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
\item \code{\link{LandReform}} : Politics and land reforms in India
}
\item \bold{ O15 } :  Economic Development: Human Resources; Human Development; Income Distribution; Migration 
\itemize{
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
\item \code{\link{IneqGrowth}} : Inequality and growth
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ O16 } :  Economic Development: Financial Markets; Saving and Capital Investment; Corporate Finance and Governance 
\itemize{
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IneqGrowth}} : Inequality and growth
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ O17 } :  Formal and Informal Sectors; Shadow Economy; Institutional Arrangements 
\itemize{
\item \code{\link{LandReform}} : Politics and land reforms in India
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
\item \bold{ O19 } :  International Linkages to Development; Role of International Organizations 
\itemize{
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing countries
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ O30 } :  Innovation; Research and Development; Technological Change; Intellectual Property Rights: General 
\itemize{
\item \code{\link{GiantsShoulders}} : Impact of institutions on cumulative research
}
\item \bold{ O31 } :  Innovation and Invention: Processes and Incentives 
\itemize{
\item \code{\link{Dialysis}} : Diffusion of haemodialysis technology
}
\item \bold{ O32 } :  Management of Technological Innovation and R&D 
\itemize{
\item \code{\link{RDSpillovers}} : Research and development spillovers data
}
\item \bold{ O33 } :  Technological Change: Choices and Consequences; Diffusion Processes 
\itemize{
\item \code{\link{RDSpillovers}} : Research and development spillovers data
}
\item \bold{ O41 } :  One, Two, and Multisector Growth Models 
\itemize{
\item \code{\link{Solow}} : Growth model
}
\item \bold{ O47 } :  Empirical Studies of Economic Growth; Aggregate Productivity; Cross-Country Output Convergence 
\itemize{
\item \code{\link{DemocracyIncome}} : The relation between democraty and income
\item \code{\link{DemocracyIncome25}} : The relation between democraty and income
\item \code{\link{FinanceGrowth}} : Financial institutions and growth
\item \code{\link{IneqGrowth}} : Inequality and growth
\item \code{\link{Reelection}} : Deficits and reelection
\item \code{\link{Solow}} : Growth model
\item \code{\link{TwinCrises}} : Costs of currency and banking crises
}
\item \bold{ Q11 } :  Agriculture: Aggregate Supply and Demand Analysis; Prices 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
}
\item \bold{ Q12 } :  Micro Analysis of Farm Firms, Farm Households, and Farm Input Markets 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
}
\item \bold{ Q15 } :  Land Ownership and Tenure; Land Reform; Land Use; Irrigation; Agriculture and Environment 
\itemize{
\item \code{\link{LandReform}} : Politics and land reforms in India
}
\item \bold{ R12 } :  Size and Spatial Distributions of Regional Economic Activity 
\itemize{
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ R23 } :  Urban, Rural, Regional, Real Estate, and Transportation Economics: Regional Migration; Regional Labor Markets; Population; Neighborhood Characteristics 
\itemize{
\item \code{\link{IncomeMigrationH}} : Income and Migration, household data
\item \code{\link{IncomeMigrationV}} : Income and Migration, village data
\item \code{\link{RegIneq}} : Interregional redistribution and inequalities
}
\item \bold{ R31 } :  Housing Supply and Markets 
\itemize{
\item \code{\link{HousePricesUS}} : House Prices data
}
\item \bold{ R41 } :  Transportation: Demand, Supply, and Congestion; Travel Time; Safety and Accidents; Transportation Noise 
\itemize{
\item \code{\link{SeatBelt}} : Seat belt usage and traffic fatalities
}
\item \bold{ Z12 } :  Cultural Economics: Religion 
\itemize{
\item \code{\link{Donors}} : Dynamics of charitable giving
}
\item \bold{ Z13 } :  Economic Sociology; Economic Anthropology; Language; Social and Economic Stratification 
\itemize{
\item \code{\link{ScrambleAfrica}} : The long-run effects of the scramble for Africa
}
}
}
\keyword{documentation}
