\name{IneqGrowth}
\docType{data}
\alias{IneqGrowth}
\title{Inequality and Growth}
\description{5-yearly observations of 266 world from 1961 to 1995

    \emph{number of observations} : 1862

    \emph{number of  time-series } :  7 

    \emph{country} :  country 

    \emph{package} : panel

    \emph{JEL codes: }O47, O15, C23, C33, O16

    \emph{Chapter} : 07

}
\usage{data(IneqGrowth)}
\format{A dataframe containing:
  \describe{
    \item{country}{country name}
    \item{period}{the period}
    \item{growth}{growth rate}
    \item{yssw}{years of secondary schooling among women, lagged}
    \item{yssm}{years of secondary schooling among men, lagged}
    \item{pinv}{price level of investment, lagged}
    \item{lgdp}{log initial gdp per capita}
    \item{gini}{gini index}
  }
}
\source{\url{http://www.cgdev.org/content/publications/detail/14256}}
\references{
Forbes, Kristin J. (2000) \dQuote{A Reassessment of the Relationship Between Inequality and Growth}, \emph{American Economic Review}, \bold{90(4)}, 869-887, \doi{10.1257/aer.90.4.869}
.

Roodman, David (2009) \dQuote{A Note on the Theme of Two Many Instruments}, \emph{Oxford Bulletin of Economics An Statistics}, \bold{71(1)}, 135--158, \doi{10.1111/j.1468-0084.2008.00542.x}
.

}
\keyword{datasets}
