\name{InflationProfit}
\docType{data}
\alias{InflationProfit}
\title{Search Costs and the Relation Between Inflation and Profit }
\description{
  yearly observations of  57  from 1978  to 1997

 \emph{number of observations} :  1140

\emph{economic topic} : macroeconomics

\emph{econometrics topic} : error component

}
\usage{data(InflationProfit)}
\format{A dataframe containing :
  \describe{
    \item{sic}{ sic code}
    \item{year}{year}
    \item{sales}{the sales}
    \item{cost}{cost of goods sold}
    \item{aginf}{aggregate inflation}
    \item{indinf}{industry inflation}
    \item{search}{ search costs}
    \item{growth}{growth rate of real gdp}
  }
}
\source{
    Journal Money, Credit and Banking data archive : \url{http://jmcb.osu.edu}.

}
\references{
     and   (2004) \dQuote{The role of search costs in determining the relationship between inflation and profit margins}, \emph{Journal of Money, Credit and Banking}, \bold{36(1)}, 139--149.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
