\name{Lakisha}
\docType{data}
\alias{Lakisha}
\title{Resumes With 'white' Or 'african-american' Names }
\description{
  a cross-section

 \emph{number of observations} :  4870

   \emph{country} :  United States

\emph{economic topic} : labor economics

\emph{econometrics topic} : binomial model

}
\usage{data(Lakisha)}
\format{A time serie containing :
  \describe{
    \item{call}{has the employer responded to the resume ?}
    \item{firstname}{a factor of 36 first names (18 for males and 18 for females, 18 for blacks and 18 for whites)}
    \item{city}{one of 'boston' and 'chicago'}
    \item{sex}{one of 'female', 'male'}
    \item{race}{one of 'black' and 'white'}
    \item{quality}{a subjective measure of the quality of the resume, one of 'low' and 'high'}
    \item{exp}{ years of experience}
    \item{volunteer}{volunteering experience ?}
    \item{military}{military experience ?}
    \item{email}{ email address ?}
    \item{empholes}{employment holes ?}
    \item{worksch}{work in school ?}
    \item{honors}{honors ?}
    \item{compsk}{computer skills ?}
    \item{specsk}{special skills ?}
    \item{apblack}{fraction african-americans in applicant's zip code}
    \item{apwhite}{fraction whites in applicant's zip code}
    \item{apdrop}{ fraction high school dropouts in applicant's zip code}
    \item{apcolp}{fraction college or more in applicant's zip code}
    \item{alinc}{log of median household income in applicant's zip code}
    \item{occupation}{the occupation, a factor with 6 levels}
    \item{sector}{the secor, a factor with 7 levels}
    \item{req}{any requirement ?}
    \item{reqexp}{experience requirement ?}
    \item{reqcom}{communication skills required ?}
    \item{reqeduc}{education requirement ?}
    \item{reqcomp}{computer skills required ?}
    \item{reqorg}{organization skills ?}
    \item{fed}{federal contractor ?}
    \item{eoe}{equal opportunity employer ?}
    \item{ownership}{ownership status ?}
    \item{epblack}{fraction african-americans in employer's zip code}
    \item{epwhite}{fraction whites in employer's zip code}
    \item{epdrop}{ fraction high school dropouts in employer's zip code}
    \item{epcolp}{fraction college or more in employer's zip code}
    \item{elinc}{log of median household income in employer's zip code}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Bertrand, Marianne  and  Sendhil  Mullainathan (2004) \dQuote{Are Emily and Greg more employale than Lakisha and Jamal? A field experiment on labor market discrimination}, \emph{American Economic Review}, \bold{94(4)}, 991-1013.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
