\name{Migrants}
\docType{data}
\alias{Migrants}
\title{Returns to Migration }
\description{
   observations of 1259 

 \emph{number of observations} :  2655

\emph{economic topic} : Labor economics

\emph{econometrics topic} : Fixed effects model

}
\usage{data(Migrants)}
\format{A dataframe containing :
  \describe{
    \item{household}{household index}
    \item{region}{one of 'urban' and 'rural'}
    \item{migrant}{'yes' for those who migrated to the United-States}
    \item{age}{the age un years}
    \item{lwage}{the logarithm of the wage}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Abramitsky, R.,  L.P.   Boutsan and  K.  Eriksson (2012) \dQuote{Europe's tired, poor, huddled masses: self-selection and economic outcomes in the age of mass migration}, \emph{American Economic Review}, \bold{102(5)}, 1832-1856.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
