\name{MoneyJapan}
\docType{data}
\alias{MoneyJapan}
\title{Money Demand of Japan }
\description{
  yearly observations of  47 prefectures from 1992  to 1997

 \emph{number of observations} :  282

   \emph{country} :  Japan

\emph{economic topic} : macroeconomics

\emph{econometrics topic} : variable coefficients

}
\usage{data(MoneyJapan)}
\format{A dataframe containing :
  \describe{
    \item{prefecture}{the prefecture}
    \item{year}{the year}
    \item{cpi}{the price deflator}
    \item{m2}{the level counterpart of M2}
    \item{income}{the income}
    \item{m1}{the level counterpart of M1}
    \item{bond}{bond rate 5 years}
  }
}
\source{
    Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.

}
\references{
    Cheng Hsiao, Yan Shen (2005) \dQuote{Aggregate vs. Disaggregate Data Analysis - A Paradox in the Estimation of Money Demand Function of Japan Under the Low Interest Rate Policy}, \emph{Journal of Applied Econometrics}, \bold{20(5)}, 579--601.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
