\name{NaturalGas}
\docType{data}
\alias{NaturalGas}
\title{Interstate Natural Gas Industry }
\description{
   observations of  24 firms from 1977  to 1985

 \emph{number of observations} :  216

   \emph{country} :  United States

\emph{economic topic} : producer behavior

\emph{econometrics topic} : error component

}
\usage{data(NaturalGas)}
\format{A dataframe containing :
  \describe{
    \item{firm}{the name of the firm}
    \item{year}{year}
    \item{q}{trillion cubic feet-miles of natural gas transported}
    \item{xl}{number of employees}
    \item{pl}{labor price}
    \item{xe}{billion cubic feet of natural gas used to fuel compressors}
    \item{pe}{energy price}
    \item{xh}{horsepower rating of all compressors}
    \item{rrh}{ex post internal rate of return per compressor horsepower}
    \item{ph}{service price per compressor horsepower}
    \item{xp}{tons of steel in transmission line pipes}
    \item{rrp}{ex post internal rate of return per ton of steel in transmission line pipes}
    \item{pp}{service price per ton of steel in transmission line pipes}
    \item{capu}{capacity utilization or transmission system}
  }
}
\source{
    Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.

}
\references{
    Robin C. Sickles and Mary L. Streitwieser, (1998) \dQuote{An Analysis of Technology, Productivity, and Regulatory Distortion in the Interstate Natural Gas Transmission Industry: 1977-1985}, \emph{Journal of Applied Econometrics}, \bold{13(4)}, 377--395.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
