\name{RDPerfComp}
\docType{data}
\alias{RDPerfComp}
\title{R and D Performing Companies}
\description{yearly observations of 509 firms from 1982 to 1989

    \emph{number of observations} : 4072

    \emph{number of  time-series } :  8 

    \emph{country} :  United States 

    \emph{package} : panel

    \emph{JEL codes: }C51, D24

    \emph{Chapter} : 07

}
\usage{data(RDPerfComp)}
\format{A dataframe containing:
  \describe{
    \item{id}{firm identifier}
    \item{year}{year}
    \item{y}{production in logs}
    \item{n}{labor in logs}
    \item{k}{capital in logs}
  }
}
\source{author's website \url{https://www.nuffield.ox.ac.uk/users/bond/index.html}}
\references{
Blundell, Richard and Stephen Bond (2000) \dQuote{GMM Estimation with Persistent Panel Data: An Application to Production Functions}, \emph{Econometric Reviews}, \bold{19(3)}, 321-340, \doi{10.1080/07474930008800475}
.

}
\keyword{datasets}
