\name{Recycling}
\docType{data}
\alias{Recycling}
\title{Social Norms and Recycling }
\description{
  a cross-section

 \emph{number of observations} :  608

   \emph{country} :  United States

\emph{economic topic} : public economics

\emph{econometrics topic} : ordered models

}
\usage{data(Recycling)}
\format{A dataframe containing :
  \describe{
    \item{bottles}{number of bottles out of ten}
    \item{green}{considers self environmentalist}
    \item{msa}{lives in a metropolitan statistical area ?}
    \item{income}{income in 10,000 dollars increments}
    \item{them}{neighbors would be upset if they saw someone trashing recyclables}
    \item{me}{respondent would be upset if they saw neighbors trashing recyclables}
    \item{region}{region, one of 'northeast', 'midwest', 'south', 'west'}
    \item{deplaw}{deposit law, one of 'non', 'plbot' (law covering plastic water bottles) and 'noplbot' (law not covering plastic water bottles)}
    \item{lawreq}{state law requirements, one of 'none', 'recplan' (requires a recycling plan) and 'recgoal' (requires a recycling goal)}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Viscusi, W. Kip,  Joel   Huber and  Jason  Bell (2011) \dQuote{Promoting recycling: private values, social norms, and economic incentives}, \emph{American Economic Review}, \bold{101(3)}, 65-70.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
