\name{RegIneq}
\docType{data}
\alias{RegIneq}
\title{Interregional Redistribution and Inequalities}
\description{yearly observations of 17 countries from 1982 to 1999

    \emph{number of observations} : 102

    \emph{number of  time-series } :  6 

    \emph{country} :  oecd 

    \emph{JEL codes: }D72, H23, H71, H73, H77, R12, R23

    \emph{Chapter} : 07

}
\usage{data(RegIneq)}
\format{A dataframe containing:
  \describe{
    \item{country}{the country}
    \item{period}{the period}
    \item{regineq}{coefficient of variatio of regional gdp per capita}
    \item{gdppc}{real gross domestic product per capita}
    \item{pop}{total population}
    \item{popgini}{gini coefficient of regional population size}
    \item{urban}{share of urban living population}
    \item{social}{total government social expenditures as share of gdp}
    \item{unempl}{unemployment rate}
    \item{dec}{sub-national expenditures as share of total government expenditures}
    \item{transrev}{grants received by national and sub-national governments from other levels of government as share of total government revenues}
    \item{transaut}{sub-national non autonomous revenues as share of total government revenues}
  }
}
\source{Review of Economic Studies' web site \url{http://restud.oxfordjournals.org/}}
\references{
Anke S. Kessler and Nico A. Hansen and Christian Lessmann (2011) \dQuote{Interregional Redistribution and Mobility in Federations: a Positive Approach}, \emph{Review of Economic Studies}, \bold{78(4)}, 1345-1378, \doi{10.1093/restud/rdr003}
.

}
\keyword{datasets}
