\name{SchoolDist}
\docType{data}
\alias{SchoolDist}
\title{School Distance, Education and Earnings }
\description{
  yearly observations of 171 farms

 \emph{number of observations} :  3010

   \emph{country} :  United States

\emph{economic topic} : labor economics

\emph{econometrics topic} : instrumental variables

}
\usage{data(SchoolDist)}
\format{A time serie containing :
  \describe{
    \item{wage}{hourly wage}
    \item{educ}{years of education}
    \item{age}{age in years}
    \item{exper}{years of experience}
    \item{famsit}{ family situation, one of 'momdad14' (lives with both parents), 'sinmom14' (lives with the mother), 'step14' (lives with step parent) or 'other'}
    \item{region}{the region, a factor with 9 levels}
    \item{black}{black}
    \item{smsa}{lives in smsa}
    \item{smsa66}{lived in smsa in 1966}
    \item{south}{lives in the south}
    \item{south66}{lived in the south in 1966}
    \item{nearc2}{2-year college proximity}
    \item{nearc4}{4-year college proximity}
    \item{iq}{intelligence quotient score}
    \item{kww}{knowledge of the world of work score}
    \item{fatheduc}{ years of education of the father}
    \item{motheduc}{years of  education of the mother}
  }
}
\source{
    Wooldridge's "Econometrics analysis of cross section and panel data" website \url{http://mitpress.mit.edu/books/econometric-analysis-cross-section-and-panel-data}.

}
\references{
    Card, David (1993) \dQuote{Using geographic variation in college proximity to estimate return to schooling}, \emph{Nber Working Paper}, \bold{4483}.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
