\name{SovereignDefault}
\docType{data}
\alias{SovereignDefault}
\title{Capital Flows, Commodities and Sovereign Defaults }
\description{
   observations of  124 countries from 1800  to 2015

 \emph{number of observations} :  26784

   \emph{country} :  world

\emph{economic topic} : financial economics

\emph{econometrics topic} : binomial variable

}
\usage{data(SovereignDefault)}
\format{A dataframe containing :
  \describe{
    \item{country}{country name}
    \item{year}{year}
    \item{indyear}{year of independence}
    \item{default}{dummy for the start of the default}
    \item{peak}{ end year of capital flow boom}
  }
}
\source{
    American Economic Association's website \url{http://www.aeaweb.org}.

}
\references{
    Reinhart, Carmen M.,  Vincent   Reinhart and  Christoph  Trebesch (2016) \dQuote{Global cycles: capital flows, commodities, and sovereign defaults, 1815-2015}, \emph{American Economic Review}, \bold{106(5)}, 574-580.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
