\name{TwinCrises}
\docType{data}
\alias{TwinCrises}
\title{Costs of Currency and Banking Crises}
\description{yearly observations of 22 countries from 1970 to 1997

    \emph{number of observations} : 616

    \emph{number of  time-series } :  28 

    \emph{country} :  

    \emph{JEL codes: }F32, G15, G21, O16, O19, O47

    \emph{Chapter} : 06

}
\usage{data(TwinCrises)}
\format{A dataframe containing:
  \describe{
    \item{country}{the country name}
    \item{year}{the year}
    \item{gdp}{real gdp growth}
    \item{pubsurp}{change in budget surplus to real gdp ratio}
    \item{credit}{credit growth}
    \item{extgdp}{external growth rates (weight average)}
    \item{exr}{real exchange rate overvaluation}
    \item{open}{openess}
    \item{curcrises}{currency crises}
    \item{bkcrises}{banking crises}
    \item{twin}{twin crises}
    \item{area}{a factor with levels 'other', 'asia' and 'latam' (for latin America)}
  }
}
\source{Journal of Money, Credit and Banking : \url{https://jmcb.osu.edu/archive/}}
\references{
Hutchison, Michael M. and Ilan Noy (2005) \dQuote{How Bad Are Twins ? Output Costs of Currency and Banking Crises}, \emph{Journal of Money, Credit and Banking}, \bold{37(4)}, 725--752.

}
\keyword{datasets}
