\name{Twins}
\docType{data}
\alias{Twins}
\title{Education and Earnings of British Twins }
\description{
   observations of 214 Twins

 \emph{number of observations} :  428

   \emph{country} :  United-Kingdown

\emph{economic topic} : Labor Economics

\emph{econometrics topic} : Fixed effects model

}
\usage{data(Twins)}
\format{A dataframe containing :
  \describe{
    \item{family}{family index}
    \item{earning}{hourly wage rate}
    \item{educ}{reported years of schooling}
    \item{educt}{estimated years of schooling}
    \item{age}{age}
    \item{londse}{lives in }
    \item{part}{part-time job ?}
    \item{married}{married ?}
    \item{tenure}{tenure in years}
  }
}
\source{
    American Economic Review data archive : \url{http://aeaweb.org/aer/}.

}
\references{
    Bonjour, D.,  L.F.   Cherkas,  J.E.   Haskel,  D.D.   Hawkes and  T.D.  Spector (2003) \dQuote{Returns to education: Evidence from U.K. twins}, \emph{American Economic Review}, \bold{93(5)}, 1799-1812.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
