\name{Twinsburg}
\docType{data}
\alias{Twinsburg}
\title{Education and Earnings of American Twins }
\description{
  a cross-section

 \emph{number of observations} :  340

   \emph{country} :  United-States

\emph{economic topic} : Labor Economics

\emph{econometrics topic} : 3SLS

}
\usage{data(Twinsburg)}
\format{A dataframe containing :
  \describe{
    \item{age}{age in years}
    \item{white}{=1 for white twins (note than one value is 0.5)}
    \item{female}{=1 for female twins (note than one value is 0.33)}
    \item{daded}{years of education of the father}
    \item{momed}{years of education of the mother}
    \item{educ1}{self reported education for the first twin}
    \item{educ2t}{education for the second twin reported by his twin}
    \item{educ1t}{education for the first twin reported by his twin}
    \item{educ2}{self reported education for the second twin}
    \item{lwage1}{ log of the hourly wage for the first twin}
    \item{uncov1}{=1 if the first twin is covered by a union}
    \item{married1}{= 1 if the first twin is married}
    \item{tenure1}{tenure in years for the first twin}
    \item{lwage2}{ log of the hourly wage for the second twin}
    \item{uncov2}{=1 if the second twin is covered by a union}
    \item{married2}{= 1 if the second twin is married}
    \item{tenure2}{tenure in years for the second twin}
  }
}
\source{
     Industrial Relations Secion, Princeton University \url{http://www.irs.princeton.edu/publications/datasets}.

}
\references{
    .

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
