\name{WomenWages}
\docType{data}
\alias{WomenWages}
\title{Wages of White Females }
\description{
  yearly observations of 579 females

 \emph{number of observations} :  6948

   \emph{country} :  United States

\emph{economic topic} : labor economics

\emph{econometrics topic} : endogeneity

}
\usage{data(WomenWages)}
\format{A dataframe containing :
  \describe{
    \item{id}{individual identifier}
    \item{year}{survey year}
    \item{age}{age in years}
    \item{educ}{years of schooling}
    \item{children}{total number of children in family unit ages 0-17}
    \item{working}{participation, \code{'yes'} or \code{'no'}}
    \item{lhwage}{log of real average hourly earnings}
    \item{lhwage80}{log earnings in 1980}
  }
}
\source{
    Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.

}
\references{
    Semykina, Anastasia  and  Jeffrey M.  Wooldridge (2012) \dQuote{Estimation of Dynamic Panel Data Models with Sample Selection}, \emph{Journal of Applied Econometrics}, \bold{forthcoming}.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
