\name{CallBacks}
\docType{data}
\alias{CallBacks}
\title{Callbacks to Job Applications}
\description{a pseudo-panel of 1518 resumes from 2014

    \emph{number of observations} : 6072

    \emph{number of  individual observations } :  4 

    \emph{country} :  United States 

    \emph{package} : binomial

    \emph{JEL codes: }E24, E32, J14, J22, J23, J64

    \emph{Chapter} : 08

}
\usage{data(CallBacks)}
\format{A dataframe containing:
  \describe{
    \item{jobid}{the job index}
    \item{unempdur}{unemployment  duration in month}
    \item{interim}{a dummy for interim experience}
    \item{callback}{a dummy for call backs}
    \item{old}{a dummy for age 57-58}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Farber, Henry S.; Silverman, Dan and Till von Wachter (2016) \dQuote{Determinants of Callbacks to Job Applications: An Audit Study}, \emph{American Economic Review}, \bold{106(5)}, 314-318, \doi{10.1257/aer.p20161010}
.

}
\keyword{datasets}
