\name{CoordFailure}
\docType{data}
\alias{CoordFailure}
\title{How to Overcome Organization Failure in Organization}
\description{a pseudo-panel of 240 individuals

    \emph{number of observations} : 7168

    \emph{number of  individual observations } :  30 

    \emph{country} :  United States and Spain 

    \emph{JEL codes: }C92, D23

    \emph{Chapter} : 08

}
\usage{data(CoordFailure)}
\format{A dataframe containing:
  \describe{
    \item{firm}{the firm index}
    \item{id}{the individual index}
    \item{period}{the period}
    \item{place}{either Cleveland or Barcelona}
    \item{bonus1}{the bonus for the first block of 10 rounds}
    \item{bonus2}{the bonus for the second block of 10 rounds}
    \item{bonus3}{the bonus for the third block of 10 rounds}
    \item{effort}{the level of effort of the employee}
  }
}
\source{American Economic Association Data Archive : \url{http://aeaweb.org/aer/}}
\references{
Brandts, Jordi and David J. Cooper (2006) \dQuote{A Change Would Do You Good... An Experimental Study on How to Overcome Coordination Failure in Organizations}, \emph{American Economic Review}, \bold{96(3)}, 669-693, \doi{10.1257/aer.96.3.669}
.

}
\keyword{datasets}
