\name{IncomeMigrationV}
\docType{data}
\alias{IncomeMigrationV}
\title{Income and Migration (village Data) }
\description{
  3-yearly observations of  44674 villages from 2005  to 2008

 \emph{number of observations} :  89348

   \emph{country} :  Indonesia

\emph{economic topic} : economic development, innovation, technological change, and growth

\emph{econometrics topic} : censored and truncated variable

}
\usage{data(IncomeMigrationV)}
\format{A dataframe containing :
  \describe{
    \item{village}{village index}
    \item{year}{the year}
    \item{emigration}{share of the emigrants in the total population}
    \item{district}{the district of the village}
    \item{price}{rice price shock}
    \item{rain}{rain shock}
    \item{pareto}{pareto parameter of the landholdings distribution}
  }
}
\source{
    American Economic Association's website \url{http://www.aeaweb.org}.

}
\references{
    Bazzi, Samuel (2017) \dQuote{Wealth heterogeneity and the income elasticity of migration}, \emph{American Economic Journal, Applied Economics}, \bold{9(2)}, 219--255.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
