\name{Index.Economics}
\docType{data}
\alias{Index.Economics}
\title{Economic fields}
\description{
\itemize{
\item International Economics 
 \itemize{
\item \code{\link{TradeEU}} : Trade in the European Union  
}
\item Labor Economics 
 \itemize{
\item \code{\link{Twins}} : Education and Earnings of British Twins  
\item \code{\link{Twinsburg}} : Education and Earnings of American Twins  
}
\item Labor economics 
 \itemize{
\item \code{\link{Migrants}} : Returns to Migration  
}
\item consumer behavior 
 \itemize{
\item \code{\link{DoctHospVisits}} : Doctor and Hospital Visits  
\item \code{\link{EnergyDemand}} : Elasticities of Energy Demand  
\item \code{\link{SmokingTrends}} : Smoking Trends  
\item \code{\link{TobaccoPanel}} : Consumption of Tobacco  
}
\item development economics 
 \itemize{
\item \code{\link{SlaveTrade}} : Slave Trade and Economic Development  
}
\item economics of crime 
 \itemize{
\item \code{\link{CarThefts}} : Police Presence and Thefts  
}
\item instrumental variables 
 \itemize{
\item \code{\link{Prison}} : Prison and Crimes  
}
\item labor economics 
 \itemize{
\item \code{\link{Adoptees}} : Adoptees Education  
\item \code{\link{Beauty}} : Beauty and Wages  
\item \code{\link{FemLaborSupply}} : Labor Supply of Females  
\item \code{\link{JobTraining}} : Job Training Program and Earnings  
\item \code{\link{Lakisha}} : Resumes With 'white' Or 'african-american' Names  
\item \code{\link{MalesWages}} : Wages and Education of Young Males  
\item \code{\link{SchoolDist}} : School Distance, Education and Earnings  
\item \code{\link{WomenWages}} : Wages of White Females  
}
\item macroeconomics 
 \itemize{
\item \code{\link{AidGrowth}} : Aid and Growth  
\item \code{\link{BlackExRate}} : Exchange Rate on the Black Markets  
\item \code{\link{FinanceGrowth}} : Financial Institutions and Growth  
\item \code{\link{ForeignTrade}} : Foreign Trade of Developing Countries  
\item \code{\link{GrowthProfitBanks}} : Dynamics of Growth and Profitability in Banking  
\item \code{\link{IneqGrowth}} : Inequality and Growth  
\item \code{\link{InflationProfit}} : Search Costs and the Relation Between Inflation and Profit  
\item \code{\link{InstabilityInflation}} : Political Instability and Inflation  
\item \code{\link{MoneyJapan}} : Money Demand of Japan  
\item \code{\link{MonthExRate}} : Monthly Exchange Rates  
\item \code{\link{OpennessInflation}} : Trade Openness and Inflation Episodes  
\item \code{\link{QuartExRate}} : Quarterly Exchange Rates  
\item \code{\link{Solow}} : Growth Model  
\item \code{\link{TobinQ}} : the Q Theory of Investment  
\item \code{\link{TradeFDI}} : Trade and Foreign Direct Investment in Germany and the United States  
\item \code{\link{TwinCrises}} : Costs of Currency and Banking Crises  
}
\item producer behavior 
 \itemize{
\item \code{\link{AirlinesExCap}} : Excess Capacity in Us Airlines  
\item \code{\link{Apples}} : Apple Production  
\item \code{\link{BankCosts}} : Costs of the Us Banking Industry  
\item \code{\link{ElectrSupply}} : Electricity Supply  
\item \code{\link{InvRate}} : Investment Rate  
\item \code{\link{LargeBanks}} : Costs of the Us Large Banks  
\item \code{\link{NaturalGas}} : Interstate Natural Gas Industry  
\item \code{\link{RDPerfComp}} : R and D Performing Companies  
\item \code{\link{RiceFarms}} : Production of Rice in India  
\item \code{\link{TexasElectr}} : Production of Electricity in Texas  
\item \code{\link{Tileries}} : Production of Tileries in Egypt  
\item \code{\link{TurkishBanks}} : Turkish Banks  
}
\item public economics 
 \itemize{
\item \code{\link{DemocracyIncome}} : the Relation Between Democraty and Income  
\item \code{\link{DemocracyIncome25}} : the Relation Between Democraty and Income  
\item \code{\link{Dialysis}} : Diffusion of Haemodialysis Technology  
\item \code{\link{Fairness}} : Perveived Fairness of Rules for Allocating Seats in Trains and Parking Spaces  
\item \code{\link{Rapes}} : Sexual Offences and Rapes Deterance  
\item \code{\link{Recycling}} : Social Norms and Recycling  
\item \code{\link{Reelection}} : Deficits and Reelection  
\item \code{\link{RegIneq}} : Interregional Redistribution and Inequalities  
\item \code{\link{Religion}} : Religion and Growth  
\item \code{\link{SeatBelt}} : Seat Belt Usage and Traffic Fatalities  
\item \code{\link{Terrorism}} : Terrorist Incidents  
\item \code{\link{VehiculeTax}} : Tax for Vehicule Registration  
}
}

}
\keyword{documentation}
