\name{IneqGrowth}
\docType{data}
\alias{IneqGrowth}
\title{Inequality and Growth }
\description{
  5-yearly observations of  266   from 1961  to 1995

 \emph{number of observations} :  1862

\emph{economic topic} : macroeconomics

\emph{econometrics topic} : dynamic model

}
\usage{data(IneqGrowth)}
\format{A dataframe containing :
  \describe{
    \item{country}{country name}
    \item{period}{the period}
    \item{growth}{growth rate}
    \item{yssw}{years of secondary schooling among women, lagged}
    \item{yssm}{years of secondary schooling among men, lagged}
    \item{pinv}{price level of investment, lagged}
    \item{lgdp}{log initial gdp per capita}
    \item{gini}{gini index}
  }
}
\source{
    \url{http://www.cgdev.org/content/publications/detail/14256}.

}
\references{
    Forbes, Kristin J. (2000) \dQuote{A Reassessment of the relationship between inequality and growth}, \emph{American Economic Association}, \bold{90(4)}, 869-887.

    Roodman, David (2009) \dQuote{A note on the theme of two many instruments}, \emph{Oxford Bulletin of Economics an Statistics}, \bold{71(1)}, 135--158.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
