\name{InvRate}
\docType{data}
\alias{InvRate}
\title{Investment Rate }
\description{
  yearly observations of  703 firms from 1987  to 2000

 \emph{number of observations} :  5669

   \emph{country} :  United States

\emph{economic topic} : producer behavior

\emph{econometrics topic} : endogeneity

}
\usage{data(InvRate)}
\format{A dataframe containing :
  \describe{
    \item{firm}{firm}
    \item{year}{year}
    \item{invrate}{investment rate}
  }
}
\source{
    author's website \url{http://www.nuff.ox.ac.uk/users/bond/}.

}
\references{
    Stephen R. Bond, (2002) \dQuote{Dynamic panel data models: a guide to micro data methods and practise}, \emph{Portugese Economic Journal}, \bold{1}, 141-162.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
