\name{RegIneq}
\docType{data}
\alias{RegIneq}
\title{Interregional Redistribution and Inequalities }
\description{
   observations of  17 countries from 1982  to 1999

 \emph{number of observations} :  102

   \emph{country} :  oecd

\emph{economic topic} : public economics

\emph{econometrics topic} : endogeneity

}
\usage{data(RegIneq)}
\format{A dataframe containing :
  \describe{
    \item{country}{the country}
    \item{period}{the period}
    \item{regineq}{coefficient of variatio of regional gdp per capita}
    \item{gdppc}{real gross domestic product per capita}
    \item{pop}{total population}
    \item{popgini}{gini coefficient of regional population size}
    \item{urban}{share of urban living population}
    \item{social}{total government social expenditures as share of gdp}
    \item{unempl}{unemployment rate}
    \item{dec}{sub-national expenditures as share of total government expenditures}
    \item{transrev}{grants received by national and sub-national governments from other levels of government as share of total government revenues}
    \item{transaut}{sub-national non autonomous revenues as share of total government revenues}
  }
}
\source{
    Review of Economic Studies' web site \url{http://restud.oxfordjournals.org/}.

}
\references{
    Anke S. Kessler and Nico A. Hansen, (2012) \dQuote{Interregional redistribution and mobility in federations: a positive approach}, \emph{Review of Economic Studies}, \bold{78(4)}, 1345-1378.

}
\seealso{\code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}

}\keyword{datasets}
